requirements.txt
Flask
easyocr
Pillow
opencv-python-headless==4.5.4.60
gunicorn
